rag/requirements.txt
# RAG Chat Interface — Phase 3 Python dependencies
# Install with: pip install -r requirements.txt

llama-index-core>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-embeddings-ollama>=0.5.0
llama-index-llms-ollama>=0.5.0
chromadb>=0.6.0
python-dotenv>=1.0.0
